requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.57.0
accelerate>=0.20.0
datasets>=2.14.0

# Vision & Image Processing
opencv-python>=4.8.0
pillow>=10.0.0

# Data Science & Utilities
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Dataset Download
kaggle>=1.5.16

# Model Optimization & Deployment
onnx>=1.14.0
onnxruntime>=1.15.0

# Logging & Monitoring (optional)
tensorboard>=2.13.0
wandb>=0.15.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
